+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 * Adding `#.count` alias to `MultiMap`.
 * Fixing bug when feeding invalid values to a `MultiSet`.
 * Fixing `.from` static methods not taking byte arrays into account.
+* Fixing bugs related to `Stack.pop` edge cases. 
 * Optimizing `Stack` performance.
 
 ## 0.17.0